Improve panic handling
The d.Panic() call should not be passed the debug.Stack(), as this could
be a performance bottleneck. If necessary, it can be called in the
d.Panic() call itself with the same result. Instead, pass the returned
value `r` of the recovery, to see the reason for the panic.

Similarly, if no panic handling is defined, print the `r` value to allow
for better debug messages.